# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Blue Sky API Docs",
    "version": "1.0.0",
    "title": "Blue Sky API",
    "termsOfService": "",
    "contact": {
      "email": "ollie.tinsley@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "http://ec2-34-219-189-208.us-west-2.compute.amazonaws.com",
  "basePath": "/api",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/users": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Get all users",
        "description": "",
        "operationId": "getUsers",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/User"
              }
            }
          },
          "404": {
            "description": "No users exist"
          },
          "500": {
            "description": "API connection error"
          }
        }
      },
      "post": {
        "tags": [
          "user"
        ],
        "summary": "Add an user",
        "description": "",
        "operationId": "createUser",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "user",
            "description": "The user to create.",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User succesfully created",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "400": {
            "description": "Missing required parameter",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "500": {
            "description": "API connection error"
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Get user by id",
        "operationId": "getUserById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of user to return",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "400": {
            "description": "User id required",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "404": {
            "description": "No user exists with id"
          },
          "500": {
            "description": "API connection error"
          }
        }
      }
    },
...
    "/users/location": {
      "put": {
        "tags": [
          "user"
        ],
        "summary": "Add map location for user",
        "operationId": "setUserLocation",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Map cordinates to add to organization",
            "required": true,
            "schema": {
              "$ref": "#/definitions/LocationCriteria"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "400": {
            "description": "Missing required parameter",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "404": {
            "description": "No user exists with id"
          },
          "500": {
            "description": "API connection error"
          }
        }
      }
    }
  },
  "definitions": {
    "Error": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        },
        "required": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "optional": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
...
    "LocationCriteria": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number"
        },
        "lat": {
          "type": "number"
        },
        "lng": {
          "type": "number"
        }
      }
    },
...
    "User": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number"
        },
        "address1": {
          "type": "string"
        },
        "address2": {
          "type": "string"
        },
        "city": {
          "type": "string"
        },
        "createDate": {
          "type": "string"
        },
        "displayName": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "firstName": {
          "type": "string"
        },
        "groupName": {
          "type": "string"
        },
        "groupId": {
          "type": "number"
        },
        "lastName": {
          "type": "string"
        },
        "lat": {
          "type": "number"
        },
        "lng": {
          "type": "number"
        },
        "middleName": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "networkId": {
          "type": "number"
        },
        "organizationName": {
          "type": "string"
        },
        "organizationId": {
          "type": "number"
        },
        "phone": {
          "type": "string"
        },
        "phoneExt": {
          "type": "string"
        },
        "state": {
          "type": "string"
        },
        "suffix": {
          "type": "string"
        },
        "updateDate": {
          "type": "string"
        },
        "url": {
          "type": "string"
        },
        "zip": {
          "type": "string"
        }
      }
    }
  },
...
}